+++ daemon/containerd/image_push.go
@@ -1,0 +1,13 @@
+package containerd
+
+import (
+	"context"
+	"io"
+
+	"github.com/docker/docker/api/types"
+)
+
+// PushImage initiates a push operation on the repository named localName.
+func (i *ImageService) PushImage(ctx context.Context, image, tag string, metaHeaders map[string][]string, authConfig *types.AuthConfig, outStream io.Writer) error {
+	panic("not implemented")
+}
